requirements/base.txt
Django==1.8.8
tg-utils==0.2.0
django-crispy-forms==1.4.0
django_compressor==2.0
# Needed if you use utils.hashmodels.ModelHashIdMixin
hashids==1.1.0
django-modeltranslation==0.11

